Fix bugs introduced by reorganization

+++ local_planner/src/nodes/tree_node.h
@@ -17,7 +17,7 @@ class TreeNode
   int last_z;
   int origin;
   int depth;
-  int yaw;
+  double yaw;
 
   TreeNode();
   TreeNode(int from, int d, geometry_msgs::Point pos);
